+++ tests/integration/test_transactions.py
@@ -8,13 +8,16 @@ from tests.integration.util import (
 
 access_token = None
 
+# NOTE: Data is only generated over the past 2 years.  Ensure that the date
+# range used for transactions/get is within 2 years old
+
 
 def setup_module(module):
     client = create_client()
     pt_response = client.Sandbox.public_token.create(
         SANDBOX_INSTITUTION, ['transactions'],
-        transactions__start_date='2016-01-01',
-        transactions__end_date='2017-01-01',
+        transactions__start_date='2018-01-01',
+        transactions__end_date='2019-01-01',
     )
     exchange_response = client.Item.public_token.exchange(
         pt_response['public_token'])
@@ -29,7 +32,7 @@ def get_transactions_with_retries(client,
                                   account_ids=None,
                                   count=None,
                                   offset=None,
-                                  num_retries=5):
+                                  num_retries=50):
     response = None
     for i in range(num_retries):
         try:
@@ -49,6 +52,11 @@ def get_transactions_with_retries(client,
     return response
 
 
+def refresh_transactions(client,
+                         access_token):
+    return client.Transactions.refresh(access_token)
+
+
 def teardown_module(module):
     client = create_client()
     client.Item.remove(access_token)
@@ -59,8 +67,8 @@ def test_get():
 
     response = get_transactions_with_retries(client,
                                              access_token,
-                                             '2016-01-01',
-                                             '2017-01-01',
+                                             '2018-01-01',
+                                             '2019-01-01',
                                              num_retries=5)
     assert response['accounts'] is not None
     assert response['transactions'] is not None
@@ -69,10 +77,10 @@ def test_get():
     account_id = response['accounts'][0]['account_id']
     response = get_transactions_with_retries(client,
                                              access_token,
-                                             '2016-01-01',
-                                             '2017-01-01',
+                                             '2018-01-01',
+                                             '2019-01-01',
                                              account_ids=[account_id],
-                                             num_retries=5)
+                                             num_retries=50)
     assert response['transactions'] is not None
 
 
@@ -80,8 +88,17 @@ def test_get_with_options():
     client = create_client()
     response = get_transactions_with_retries(client,
                                              access_token,
-                                             '2016-01-01',
-                                             '2017-01-01',
+                                             '2018-01-01',
+                                             '2019-01-01',
                                              count=2,
                                              offset=1)
     assert len(response['transactions']) == 2
+
+
+def test_refresh():
+    client = create_client()
+
+    response = refresh_transactions(client,
+                                    access_token)
+    # response should be empty (aside from request_id)
+    assert response is not None